+++ expenser-ui/src/app/login.tsx
@@ -10,44 +10,28 @@ import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { apiRequest } from "@/util/apiRequest";
 import { AlertCircle, Eye, EyeOff } from "lucide-react";
-import { useState } from "react";
+import { FormEvent, useState } from "react";
 import { Link, useNavigate } from "react-router";
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
 
-const validatePassword = (pass: string) => {
-  if (pass.length < 8) {
-    return "Password must be at least 8 characters long";
-  }
-  if (!/[A-Z]/.test(pass)) {
-    return "Password must contain at least one uppercase letter";
-  }
-  if (!/[0-9]/.test(pass)) {
-    return "Password must contain at least one number";
-  }
-  return "";
-};
-
-export function SignupForm() {
+function LoginForm() {
   let navigate = useNavigate();
   const [username, setUsername] = useState("");
-  const [fullname, setFullname] = useState("");
   const [password, setPassword] = useState("");
+
   const [showPassword, setShowPassword] = useState(false);
   const togglePasswordVisibility = () => setShowPassword(!showPassword);
+
   const [error, setError] = useState("");
 
-  const handleSubmit = (e: React.FormEvent) => {
+  const handleLogin = (e: FormEvent) => {
     e.preventDefault();
-    if (!fullname) {
-      setError("Full name is required");
+    if (!username || !password) {
+      setError("Username and password are required");
       return;
     }
-    const passwordError = validatePassword(password);
-    if (passwordError) {
-      setError(passwordError);
-      return;
-    }
-    apiRequest("/cxf/user", "POST", { name: fullname, username, password })
+
+    apiRequest("/cxf/login", "POST", { username, password })
       .then(async (response) => {
         if (!response.ok) {
           const data = await response.json();
@@ -56,7 +40,9 @@ export function SignupForm() {
         return response.json();
       })
       .then(() => {
-        navigate("/auth/login");
+        // handleLogin(data.token, data.name);
+        // setContextName(data.name);
+        navigate("/");
       })
       .catch((error) => {
         setError(error.message);
@@ -66,9 +52,9 @@ export function SignupForm() {
   return (
     <Card className="mx-auto max-w-sm">
       <CardHeader>
-        <CardTitle className="text-2xl">Sign up</CardTitle>
+        <CardTitle className="text-2xl">Login</CardTitle>
         <CardDescription>
-          Provide your details below to create an account
+          Enter your email below to login to your account
         </CardDescription>
       </CardHeader>
       <CardContent>
@@ -80,16 +66,6 @@ export function SignupForm() {
               <AlertDescription>{error}</AlertDescription>
             </Alert>
           )}
-          <div className="grid gap-2">
-            <Label htmlFor="fullname">Full name</Label>
-            <Input
-              id="fullname"
-              type="fullname"
-              required
-              autoComplete="fullname"
-              onChange={(e) => setFullname(e.target.value)}
-            />
-          </div>
           <div className="grid gap-2">
             <Label htmlFor="email">Email</Label>
             <Input
@@ -133,17 +109,25 @@ export function SignupForm() {
               </Button>
             </div>
           </div>
-          <Button type="submit" className="w-full" onClick={handleSubmit}>
-            Sign up
+          <Button type="submit" className="w-full" onClick={handleLogin}>
+            Login
           </Button>
         </div>
         <div className="mt-4 text-center text-sm">
-          Already have an account?
-          <Link to="/auth/login" className="underline">
-            Login
+          Don&apos;t have an account?{" "}
+          <Link to="/auth/signup" className="underline">
+            Sign up
           </Link>
         </div>
       </CardContent>
     </Card>
+  );
+}
+
+export function LoginPage() {
+  return (
+    <div className="flex h-screen w-full items-center justify-center px-4">
+      <LoginForm />
+    </div>
   );
 }
